requirements.txt
asgiref==3.9.1
Django==5.2.5
django-allauth==65.11.1
django-enumfields==2.1.1
sqlparse==0.5.3
tzdata==2025.2
pyjwt==2.4.0
requests==2.28.1
cryptography
Pillow
django-environ